refactor(ssi): clean up controller
