+++ dva-acapy-controller/src/dva_acapy_controller/config.py
@@ -1,13 +1,13 @@
-import os
+from os import environ as env
 
 
-ADMIN_URL = os.getenv("ADMIN_URL")
-ADMIN_LABEL = os.getenv("ADMIN_LABEL")
+ADMIN_URL = env.get("ADMIN_URL")
+ADMIN_LABEL = env.get("ADMIN_LABEL")
 
-PEER_AGENT_URL = os.getenv("PEER_AGENT_URL")
-PEER_CONTROLLER_URL = os.getenv("PEER_CONTROLLER_URL")
-PEER_CONTROLLER_PORT_IN = os.getenv("PEER_CONTROLLER_PORT_IN")
-PEER_CONTROLLER_PORT_OUT = os.getenv("PEER_CONTROLLER_PORT_OUT")
-PEER_LABEL = os.getenv("PEER_LABEL")
+PEER_AGENT_URL = env.get("PEER_AGENT_URL")
+PEER_CONTROLLER_URL = env.get("PEER_CONTROLLER_URL")
+PEER_CONTROLLER_PORT_IN = env.get("PEER_CONTROLLER_PORT_IN")
+PEER_CONTROLLER_PORT_OUT = env.get("PEER_CONTROLLER_PORT_OUT")
+PEER_LABEL = env.get("PEER_LABEL")
 
 LOG_FILE = "log.json"
